backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
Werkzeug==3.1.3
pdfplumber==0.11.4
xhtml2pdf==0.2.17
beautifulsoup4==4.12.3
lxml==5.3.0
gunicorn==21.2.0
